tsdb: faster PostingsForMatchers
Upstream https://github.com/prometheus/prometheus/pull/11809 removed the
requirement for strings passed to `Postings()` to be in order.
This means we don't need to sort before calling, and can skip some
copying in the simplest case in `inversePostingsForMatcher()`.

Unfortunately we have to add one extra case where the matcher says
`""==""`, because that is used in tests to mean "fetch everything".